refactored and fixed instances of old Profile.java in to RegisteredAppOwner.java, created new different Profile.java from ProfileManager inner class
